Add a join table to persist that a change has been reviewed
When there are several moderators reviewing changes, it can be tricky
to keep track of what changes have been reviewed. This change introduces
a new model so that everytime a user reviews a change, this information
is saved in an association table that links the reviewer with the
changes and the politician who received those changes.