Update recent changes template to make tests more clear
This is a cleanup change in preparation for the next commit. The
template showing recent changes was updated to make the table
more semantic, separating the table head from the table body.
This also allows to make the current test more clear, since it is
creating two actions but asserting that there are three rows in the
table. The name of the test was also updated to something more
descriptive.