Save a row in ChangeReview when reviewer marks as reviewed
This change implements a button that can be used to send a POST
request to the database so that a row is added in the
ChangeReviewed association table. A row in this table means that
the change has been reviewed.

The button is wrapped in a form that adds hidden and required
inputs to hold the ChangeReviewed foreign keys.

This also implements the view in charge of receiving the form and
updating the aforementioned table. When the button to review the
changes is pressed, this new view will add a new row in the
ChangeReview table, using the ids retrieved from the hidden
inputs of the form and the user of the request.

In the future, group permission restrictions will be added so that
only the relevant users can mark a change as reviewed.