Show the recent pages page only to users who have review permissions
Both RecentChangesView and MarkChangeAsReviewedView where added a
required_group_name = TRUSTED_TO_REVIEW_CHANGES_GROUP_NAME so that
only users with review permissions can load the recent changes page
or submit a change to be marked as reviewed.

This will make loading the recent changes page or posting of data
to mark-change-as-reviewed to return a 403 - Forbidden error if the
user has no review permissions.