Comment out links to uk_results from post pages and the header
This so that enthusiastic people don't just click through and start
adding election results in the uk_results app for the 2017 General
Election; the ways this could interact with the older results recording
mechanism are complex and might well cause unexpected behaviour. To be
careful, we're removing these links until the results are in and we
have more time to test this and update the code (e.g. to stop duplicate
ResultEvent objects from being created, and handle corrections to
previously declared results.)