Allow a trailing slash on update links
WhoCanIVoteFor is linking to the update URL but with an extra slash on
the end, which means it's treated as a slug for the person. I don't
think this is intentional, and generally it's bad for the site to be
picky about the presence / absence of trailing slashes on URLs.