Add option to disable editing elections in Django's admin interface
In the UK the source of elections is a different service that should
be thought of as the canonical source of truth for elections.

Having the possibility that someone might edit an election in YNR (
for example to mark an election as no longer current) might cause
confusion especially if these settings get overwritten by a script.

This change makes it possible to disable the admin interface for the
Election model.