Move from Area look up to Postcode lookup
This changes the function of the home page postcode look up from
redircting to a URL with the slug of the elections and posts that were
active at the time of look up to a URL that contains the postcode,
meaning that the URL will always return currently active posts and
elections, rather than ones active at the initial look up time.

This is more in keeping with the design of all the other postcode lookup
forms we have in other places.