Fix Mayor/PCC election imports
Usually we would import the ballot id as a `Post` object
and its' first parent group as an `Election` object.

We've got 2 problems here:

1. There's no ballot id here - the organisation group is the 'leaf'
   node, so we need to import `mayor.area.date` (which is an
   organisation group) as a `Post` object.
2. We can't import `mayor.date` as an `Election` because it has no
   organisation so we need to import `mayor.area.date` again
   as its' own parent `Election` object.

To do this, we need to create a circular reference so that
`mayor.area.date` is its own parent and allow an `Election`
object to be created from a record with no children.