Use same coverage version as coveralls package
When using tox, coverage.py is installed in the tox virtualenv but the
coveralls package is installed in a different "gobal" env.

Because of version mismatches, coveralls was unable to parse the older
format.

Note that the requirements line looks like it's downgrading coverage,
but this is the exact line that's in the coveralls setup file, meaning
it should always install the same version as coveralls. Coveralls don't
seem to pin their versions to a version of coverage.py