Don't set the account adaptor by default
This code comes with the ability to prevent new account creation,
however by default we want new accounts to be made.

This moves the no-op account adaptor to the example local settings for
anyone who might want to disable account, but ensures they're enabled by
default.

Thanks to @rosewhiffen for reporting.