Basic prevention of self-locking
Of course there are other ways to prevent this that are out of scope of
this PR.

For example, this doesn't cover the case where a user with locking
permission has added all the candidates themselves but hasn't suggested
locking. This case is harder to detect.

For the time being, this should catch the most obvious cases of
self-locking.

Closes #312