Add celery
There are a couple of reasons for this:

1. Before this change, updating the search index was a blocking
operation when modifying a Person object. This would be find normally,
but slowed everything down a lot when bulk adding py SOPN or Party. It
makes sense move this work to an async task.

2. There may well be other uses for an async queue in future, so adding
this capibility to the project makes sence.

See the note in testing.py abotu running tests.