Only check creation allowed when updating
`check_creation_allowed` is confusingly named, as it really checks that
the creation of a membership is allowed, not the creation of a person is
allowed.

Because of this, we should only check this is allowed when updating a
person to add a membership.

In reality, this will be the same, as add_person and update_person are
called one after the other, both in a transaction.