Remove council control feature
This was added in a rush for work with LGiU.

It was never very good.

It was made before we had EveryElection or election IDs. It would be
better to make this again using the latest tooling we have, so it's
better to delete it.

We can back up the data and archive it somewhere before we deploy and
migrate it.