Remove the bulk of uk_results
This change is so large that it might have almost been worth deleting
the app and starting again. Taking this approach in order to save the
data we have.

The previous model allowed `n` results sets to be saved, with an admin
picking one to be 'confirmed'. Only 'confirmed' results were shown.

This turned out to be the wrong model for various reasons, so this
change moves to a single ResultSet per ballot paper (post_election or
PostExtraElection).

Most of the views and forms make no sense any more, so this commit also
removes them all, for future commits to re-add them.