Move django-popolo in to local code
This is part of the move away from using django-popolo. In order to
migrate away, it will be useful to be able to make changes or move
classes about as if it were a local package.

Once the migration is done, it will be removed from the project all
together.