Add classed for SimplePopoloFields to django settings
These classes copy the SimplePopoloField model and are meant as a short
term replacement of the database versions.

The next change is to move to declarative fields on models and forms.

Making this change allows us to move away from configuring the site in
data migrations, freeing us up to delete / reset them.