Move to using DefaultStorage for saving CSV files
This change:

1. Removes the ability/requirement to specify a fill path with prefix
for te output CSVs. Now the MEDIA_ROOT is used and the prefix is
  hard-coded
2. Moves to using DefaultStorage for writing the CSV files, meaning that
the storage backend isn't limited to the local file system.

One problem with this is that it introduces a consistency issue when the
local file system is used and the CSV file is being written. Before this
was managed by renaming a tmp file, however not all storage backends
support this, mainly S3.

The problem doesn't exist with S3 however, and that's the currently used
backend, so the problem can be mostly ignored.