Update candidates_cache_api_to_directory to use DefaultStorage
Previously this command expected to be run on a local file system with a
file system that supported things like symlinks.

This meant that the stroage backend couldn't be changed to, for example,
S3.

This set of changes moves to using DefaultStorage.

Note that this means we can't use things like shutil.rmtree. We have to
loop over files and delete them.