Refactor get_max_winners to take from PostExtraElection or 0
If we don't have a value set for the PEE then there is a high chance
that we don't know the winner count. Guessing from the election type is
likely to cause problems, so it's better to fix the data if it's
missing.

There are real cases where we don't know the winner count, and in those
cases we shouldn't show things like the "record winner" button.