Remove popolo Image field from Person
This is a URLField that's a legacy from the pre-django days.

Some people in the database have a value in this field, but they all
point to the old popit instance that hosted the first YourNextMP data.

All this data has since been moved to the Django version of popolo, and
the images that are at the URLs are imported in to the Images model.

The field isn't used anywhere, apart from a couple of tests.