Only import person images from live site
This change removes the ability to download all images from the site,
and limits it to person images only. In reality this means party emblems
don't get mirrored, but the canonical source of party images is The
Electoral Commission, and these get imported in the EC importer.

This change also stops us using the `images` api endpoint, and uses the
candidate images list directly