Remove ImageViewSet
This isn't needed any more since we removed the concept of a generic
Image model with a generic relation to other models.

The image endpoint ended up pointing to images on either the Person or
Organizations model.

Both of these have been replaced with a different model:

* Images pointing to People have gone to PersonImage
* Images pointing to Organizations have gone to PartyEmblem (Organizations with images were all parties)

This is a form of breaking change to the v0.9 API, as it removed the
endpoint, but it's less useful now anyway as the content isn't updated
any more.

The only known user of this endpoint was this project's live site
importer, and that's been updated to use the new endpoints. The logs
don't show any other use