Add on_delete to all FKs
This stops a warning about Django 2.

I hope to have squashed most of these before we get to Django 2, but
they are causing _a lot_ of warnings. It's frustrating that all
migrations have to be changed, but there we have it.

We might not want `models.CASCADE` on every FK, but as it's the default,
this doesn't change anything from where we currently are.