Ignore more files to speed up pytest collection
pytest includes loads of files when it's collecting tests. This slows
tests down by up to 10 seconds, as it's scanning all of foundation and
the download cache for something that looks like a python test.

For some reason it's also scanning _every_ `__pycache__` folder.

This change adds a list of directories we know never to contain tests,
and speeds up collection a lot