Remove Area and AreaExtra models
The functional use of this was dropped a while ago, leaving a load of
dead code and tables.

This commit removes the left over code.

In terms of API changes: we've not been populating the Area models for
while, so the CSV and API output wont change, but some keys will be
removed. I don't think this is a problem even for the old API, as no
working coce can be using them anyway.