Use int when making people with PersonFactory
This is slightly confusing, as clearly this has worked in the past,
however when moving the People model to the Person app the tests started
raising a load of inequality errors because one PK was a string and the
other was an int. Why/how the factory is making models with string PKs,
I don't know. Maybe we're using the memory version of the class and not
the one fetched from the DB.

It could be that this is unrelated to these changes, but either way I
expect it's better to use ints for PKs here