Remove Dateframeable mixin
This remove the start and end date form the Person model.

We don't use these fields anywhere.

They do have some content on the production database. Looking in to
this, it's all unneeded due to the way it was added.

In the past a signal populated this field from the birth and death
dates, but if either were removed, the signal never removed the values
from the start and end columns.

A total of 26 rows have this pattern, all of them are because of edits
to remove the birth/death dates