Add Enum for PersonIdentifier Fields
This is similar to the work to remove SimplePopoloFields from the
database layer. The values here are all the values we want in the
initial version of the "n-links" on the person page model.

One day we want to enable any type of link, specified by the user.
That's too big a change to do all at once, so step one is to remove
ComplexPopoloFields in the database but still use a fixed list of
fields.