Add unique constrains on (person,value_type).
In future we want to support any number of links, maybe any number of
links by type (A person might have more than one home page, Twitter,
etc).

For the time being, we are limited by how we're storing versions on
Person. The current format only allows a single value per field type.

Let's get the database to constrain ourselves to that too, just in case
we end up throwing data away when we revert etc.

We should remove this once we can store the data in a more ad-hoc way