Add ProcessInlineFormsMixin
This mixin helps with views where we have inline formsets and a form. It
DRYs up the job of adding the forms and processing them (including
validation)

This isn't the best place for it to live (it might even be useful as a
package), but it's staying here until a sort out of all the code in this
project that's just all over the place.