Raise expected queries
I've not spend a lot of time looking in to why deleting code has made
this go up, but I think it's to do with the way checkpoints are being
called again.

The number will change again soon, as a lot of the queries are related
to generic fields on the Person model. Let's not worry about this yet.