Add ElectionIDSwitcher
This is a class designed to be used in a URL config that returns a
different view depending on the database returning a PostExtraElection
record.

It's needed because the logical URL structure for our elecion IDs means
a single URL pattern should match either an "Election" or a
"PostExtraElection" (Soon to be Ballot).