Convert TwitterBot to use PersonIdentifiers model
This was "failing" due to using the old models rather than the new ones.
The data was being updated, but ignored in the database.

This would have been detected (and other bugs that might still exist) by
removing the `contact_details` and `identifiers` properties of the
Person model (and related models). All the data in those models exists
in the PersonIdentifier model, and they were only kept as insurance that
the PersonIdentifier models broke something.