Allow deleting PersonIdentifiers by removing the value
This will remove the whole PersonIdentifier model for this form,
including the internal_identifier. As there is no way for the user to
add an internal ID then this is the right thing to do.

If the user can't delete the PI model then things like TwitterBot might
re-create the user-deleted values at a later date.