Skip if person not found for membership
This addresses a frustrating problem where the cached person endpoints
are out of sync from the memberships endpoint (a new person was added
since the cached endpoints were last made).

In most cases it's ok to ignore the latest changes to the site, so it's
safe to just skip over memberships where a person can't be found for it.