Record a new version for each person with a legacy ID
This will cause a new "fake" version to be created, meaning that the
next time this person is edited it won't show removed identifiers.
Showing the IDs as part of another edit would be confusing to users and
make it harder to work out what actually changed in the edit.