Do away with additional_merge_actions
This was a feature that would have been handy when the code base was
supporting more than one country or more than one install with optional
models and apps installed.

This is no longer the case, so supporting a "plug in" style system for a
single install makes no sense.