Move merge person feature to class
This pulls the merging functionality out from a mix of view code and the
revert code in to it's own class that can be tested without needing
loads of set up and interacting over HTTP.

The move uncovered a few bugs in the existing code, for example it would
previously delete some logged actions when merging (#758).

This class also prevents deleting any model with existing related
objects. This should catch errors seen in #118.

It also uses `update_or_create` on OtherNames, fixing #525.