Update test data so as not to test invalid data
Previously the test set up the following:

* Tessa Jowell, parl.2015, Dulwich
* Tessa Jowell, parl.2010, Dulwich
* Shane Collins, parl.2010, Dulwich
* Shane Collins, parl.2015, Edinburgh

These 4 memberships would get squashed down to 2 memberships, because
the code had an implicit unique constraint on (membership, election).

This would cause a CASCADE DELETE on the old memberships, without any
warning (because the warning was disabled).

This is a) a possible cause of lost data (see #118) and b) not very
python-y in that it would be better to raise when this sort of integrity
error was found.

In the real world it's hard to imagine that a duplicate person would be
created in the same election, but if it did happen it's better to raise
an exception, log it and warn the user. In this case an admin should
deal with the merge manually, as something is bound to have gone really
wrong. This is annoying, but much better than just deleting a load of
data without any logging.