Pin pytest at 4.1.1
It looks like https://github.com/pytest-dev/pytest/pull/4091 broke the
setUpTestData method, meaning all fixture data was being inserted for
each test method.

This is not a fix. If pytest doesn't provide a fix, we'll have to look
at using it's session fixtures, or moving tests to full pytest functions
– this would be good in the long term (it provides nicer debugging and
less boilerplate), but would be a huge job.