Property on Ballot to get SOPN
This assumes it's only useful to see the latest SOPN, not all related to
the ballot. We do have more than one in some cases, normally because the
source updates them, fixing mistakes etc. We keep the old ones, but for
locking etc we want to look at the latest one