Link to previous candidacies' SOPNs on review page
When you're checking if a candidate you've just added is the same as one
with a similar name who stood previously, you can gain a lot of
confidence in saying "Yes, they're the same!" if they have the
same home address on the current and earlier SOPN. However, it was
previously an involved process to find the previous SOPN - you'd have
to:

 1. Open the link to the person page in a new tab
 2. Find the previous candidacy on that page, and click through to that
    ballot
 3. Click on the link to the SOPN on the ballot page

This commit saves this clicking by adding a link to the earlier SOPN, if
we have one, beside the listed candidacies on the review page. Hopefully
this will save people a lot of time when they're checking if two people
are the same.

Note that this adds a query per possible matching person who would be
shown on the review page, since the format_value takes a single Person
returned from Haystack. It would need a fair bit of rewriting of
bulk_adding/forms.py to avoid this, though. (The number of queries is
already O(n) where n is the number of possible matching people.)