Add stripped down ballot view
This is the very basics, with a new load of HTML and view code, for
showing candidates for a ballot.

A lot of features are missing in this commit, but it's too complex to
replace them all in one go.

The HTML might be best ignored as we're about to change the design of
this page anyway. The view code might be best looked at as a new view,
rather than looking at the diff.