MVP API docs in API classes
I've focused on the `Party` app for this PR, just to crack through a set
of test cases for this method.

Clearly a lot more to do, but this proves that we can actually write
docs.

There are some ugly bits, the worst is `DefaultPartyEmblemSerializer`
where I've sub-classed `PartyEmblemSerializer` just to alter the docs. I
couldn't find any other way of doing this.