Set CELERY_TASK_ALWAYS_EAGER by default
Even though the docs recommend against this[1], setting `task_always_eager`[2]
by default means that local developers and Travis don't need celery working
in order to run the code and tests.

We can set this to False in prod, to make sure celery is actually run, but
it's still a reasonable default, as the only implication is that tasks end
up blocking on requests, but still get run.

1: https://docs.celeryproject.org/en/latest/userguide/testing.html
2: https://docs.celeryproject.org/en/latest/userguide/configuration.html#std:setting-task_always_eager