Fix "Enter" title-casing the name when editing any field
A <button> with no type is treated the same as a button with its type
set to submit, and the form's default button is considered to be the
first submit button in the form. (More details and links to the spec
here: https://stackoverflow.com/q/4763638/223092 )

This meant that if you pressed "Enter" in any field of the person form,
it would title-case the name, as @sjorford reported.

This commit fixes that surprising behaviour by adding a type="button"
attribute to the <button> element. Now "Enter" anywhere in the person
edit form tries to submit the whole form, which is standard behaviour in
a form.

Partial fix for #1042