Guarantee uniqueness of manually uploaded pictures of candidates
Since the Django DEFAULT_FILE_STORAGE setting changed in production [1]
to be based on ynr.s3_storage.MediaStorage (with AWS_S3_FILE_OVERWRITE
still at its default of True) uploaded images might overwrite each
other, if they're uploaded on the same day and the with the same
supplied filename. (This has in fact happened on 91 occasions, with
consequent data loss.)

This commit changes the upload_to value on QueuedImage's image field to
be a function that generates a guaranteed unique filename, since it
incorporates a UUID. @symroe said that he prefers a fix like this
instead of setting AWS_S3_FILE_OVERWRITE = False, since apparently some
other parts of the codebase now rely on the overwrite behaviour.

[1] https://github.com/DemocracyClub/YNR_deploy/commit/7ae5b0e1aa1a69b5b470403f9b831006c49e911c

Partial fix for #1096