allow_if_trusted_to_lock on user_can_edit_membership check
This is a partial fix for #991, in that it makes the behaviour of
`user_can_edit_membership` changeable depending on the value of
`allow_if_trusted_to_lock`.

This is useful on the ballot page, as a few buttons and some other logic
is looking for this flag, and it's being set to True in some cases that
make everything cluttered and unhelpful.

The bigger fix of altering the person edit forms is still to do, but
this is good enough for the time being, and for the ballot view.