Update DC slugs when importing posts
This deals with the case that we have a post with a DC slug that then
gets a GSS code. We use (slug, org_slug, start_date) to look up posts,
and then either replace or add the identifier.

If replacing, we store the old value on a new PostIdentifier model for
later look ups and mapping between IDs