Remove need for migrations in tests
This is a yak shave that I've wanted to do for a while – we have so many
migrations that running tests without a cached database takes forever.

This stops migrations being run at all in tests.

The implications of this are that we can't use data migrations to set up
test fixtures or things like, that, but I expect this is a good idea in
general.