Order previous ballot correctly
Fixes #1182

This is a bit complex.

`get_previous_ballot_for_post` worked previously, returning the 2017
ballot when called with 2019 ballots. Two things happened here:

1. The ordering on `election__election_date` omitted the `-`, meaning it
was getting the _oldest_ ballot for the post, not the previous one

2. Because the post IDs changed internally between 2017 and 2015, we
were actually always returning the previous ballot, but by chance!

In https://github.com/DemocracyClub/yournextrepresentative/pull/1168 I
changed the post IDs, meaning `get_previous_ballot_for_post` started
returning the 2010 ballots (the oldest general election we have in the
database).

Because we don't have results for 2010, the data was omitted from the
CSV.

This single character commit fixes it, and also makes the ratio of
code change to commit message quite funny :)