Remember if a page has previously been matched
This is a small step towards preventing false positives: remember if a
page has previously been matched to a ward.

This means that if we match "broadheath" and then later try to find
"heath" in the same PDF, we'll never try to match it against the page
containing "broadheath".

This has the other advantage of being able to list unlatched pages, that
might come in handy for debugging later.