Explicitly list fields to merge
Change the merge class to list the mapping between fields and relations
on the model to the method that merges them.

This is to make it easier to detect when a field isn't caught by a
method.

To prove this, add a test that will fail if a relation or field on a
model isn't listed in the SUPPORTED_FIELDS dict