Remove ballot_guess from LoggedAction
Ballot guess was a thing we had to implement when moving away from
running the site for a single election (parl.2015) to yearly elections.

The code made the assumption that a post would be the best thing to log
an action against, when actually it's more useful for us to log against
a ballot.

We've migrated all models to have a ballot (where it should) now, so
this guess isn't useful at all.