Use prepare_value for Ballot input
This bug was caused by the form using the `to_python` value of the
field, in other words the model object. The string version of this was
something like `<Ballot
ballot_paper_id='local.leeds.morley-south.2021-05-06' winner_count=1>`.

This in turn was being submitted to the field, and a ValidationError was
being raised, as that string isn't a valid ballot_paper_id.

The fix is to use prepare_value. This method is undocumented, but is
used by Djanog form fields themselves.