Never import referendums in the EEImporter
Referendums should not be imported as there are no candidates.

Uses the exclude_election_id_regex filter in EE to exclude elections
with an id that starts with "ref.". Updated the associated tests and
fixtures to add this to the querystring of the URL used, as it will be
included in every call to the EE elections endpoint by our EEImporter.

If this requirement changes, and we decided to import them in YNR, we
could update the corresponding management command to add an option to
pass election types to be excluded in the import process