Import modified from EE, add recently_updated option
Update importer so that ee_modified is taken from the API response
from EE.

This then allows the use of last_updated option in the EE management
command, which takes the latest ee_modified value from the DB and uses
it to limit the objects retrieved from EE importer

Note: .get("modified") is used because for now the ee_modified field
allows null=True. If this changes then the existing tests will break
unless the test fixtures are updated to add 'modified' dates.