delete_deleted_elections uses --recently-updated
Method to delete deleted elections now uses the same timestamp that is
used to import changes if the recently-updated flag is used. This will
allow us to soft delete any election in EE and have this replicated in
YNR. Currently only deleted elections with a poll open date in the
future or within the last 30 days are picked up by the importer.