Update API response whe last_updated used
When a request is made using last_updated filter, we:
- Annotate the last_updated filter to the ballot queryset and order
the objects by this
- Filter on this field using the timestamp given
- Remove pagination, and instead return a maximum of 200 objects at a
time to force clients to handle large updates of ballots in chunks.
This is to protect the server from becoming overwhelmed when a large
number of ballots have been updated at the same time e.g. through a
data migration