Fix 500 error on person update when party unselected
- Fixes a bug that was causing a 500 error to be raised if a user
selected a party in the "New Ballot" field, and then unselected it.
- The party selected was still passed to the request POST dict after
the deselect button was clicked. This updates the javascript to listen
for the clear event and clears the input value.
- Also adds validation to ensure that if the form is submitted with
just a party selected and no ballot, a form validation error is
displayed rather than a 500 server error.