Fixes bug where Welsh description assigned for 'The Green Party'
Special case when Green Party candidates appear on SOPN under
'The Green Party'. There is no matching PartyDescription for this but
previously it was incorrectly matched with a Welsh party description
that startswith 'The Green Party'. This meant that wombles would have
to change the party manually to simply 'The Green Party'.
So this change removes the 'The' only when the description is exactly
'The Green Party' appears. This makes sure that if the welsh
description (or other) should be matched, it will be.