Fixes bug where Welsh description assigned for Green Party
Special case when Green Party candidates appear on SOPN under
'The Green Party' or 'The Green Party candidate'. There is no matching
PartyDescription for this but previously it was incorrectly matched
with a Welsh party description that startswith 'The Green Party'. This
meant that wombles would have to change the party manually to simply
'The Green Party'.
Therefore this adds a list of the known special cases to ensure that
they are matched simply to the main 'Green Party' object (ec_id=PP63)
but also makes sure that if the welsh description (or other) should be
matched it will be.