Change how we guess if "all" pages should be used
This makes us more cautious about how and when we match pages. This
was a bug that occured with https://candidates.democracyclub.org.uk/elections/local.city-of-london.2022-03-24/
where only the first page was matched meaning that candidartes were
missed.
Now, as well as returning "all" for a single page SOPN, also do so
when there are fewer than 4 pages to the document and there is only
a single unmatched document. From running the 'make-sopn' tests this
does not negatively affect any of the ballots in the test set, but
has resolved the issue with the city-of-london ballots