Dont allow lock suggestions to be created if ballot is locked
Fixes a bug that could occur in the following scenario:
- User A completes bulk add form to add candidates but does not
submit the final step to lock candidates
- User B also submits a lock suggestion
- User C confirms user B's lock suggestion, locking the candidates
- User A belatedly submits their lock suggestion, which would still
create a lock suggestion for the ballot and add it to the queue. But
now the "Lock ballot" button on the lock review screen would actually
unlock the ballot

These changes add additional validation to block the lock suggestion
from being created. As a safeguard, it also excludes locked ballots
from the queryset used on the lock suggestion review list.