Use simple dict, not English
Using 'english' automatically removes 'stop words' like "the" and "an".
This is normally useful when searching, but loads of English names also
contain stop words.

We've had bug reports where the name "will" isn't returned, and this is
because "will" is also a common English word.

Using 'simple' means no stop-words are removed at index time or search
time.