Redesign search until
The previous system was very slow due to a bug in postgres that made
ts_rewrite take over 9 seconds to complete a search.

This moves to using ts_rewrite to built the tsquery, but not inline. For
some reason this speeds things up lot.

Also, filter by rank and annotate party name