Filter people/ballots last_updated with gte
Avoids a potential situation where multiple objects have the same
last_updated timestamp, leading to updates being missed when
importing from YNR to WCIVF. This could be a problem particularly if
instances are updated in bulk. We have also seen some photo updates
being missed recently in WCIVF which may or may not be related.