Use iterator to stop candidates cron being killed
Since adding in previous_party_affiliations to the candidates CSV the
cron fails on the prod server. This uses the iterator queryset method
to reduce the number of objects loaded in to memory when building up
the CSV data. From manual testing on the prod server, the management
commmand completes successfully.