Don't JSON loads existing JSON
This is a long standing bug that's only just come to light.

Since https://github.com/DemocracyClub/yournextrepresentative/pull/1299
`Person.versions` has been a dative Python object. This means that we
were passing a list to `json.loads()` rather than a string as it's
expecting.

Because it's already a python object, there's no need to do anything
apart from return it.