Prevent adding candidates on ballot if we have a SOPN
This is to prevent the case where well meaning users add all the
candidates to a ballot via the "add candidate form" rather than the bulk
adding process. This isn't helpful as we need users to de-duplicate and
suggest locking.

Users who can lock ballots are able to edit candidates still. This is
because sometimes it's useful to be able to manually correct bulk adding
errors this way.