Exclude other content types from name search trigger
The OtherName model is a Django generic relatable, meaning we have a
single table with content that relates to >1 model.

Previous to this change, we were getting other names by joining on the
object_id of the OtherName model, however it was possible for content
form other models to be included in this.

For example, we used to use OtherName for parties (no longer, but the
data is still in the table on Prod), so we could have:

Person 1 -> OtherName = "John Smith"
Party 1 -> OtherName = "Cornish Independents"

The resulting name vector field would contain "john cornish smith
independents". Therefore a search for "John Cornish" would return "John
Smith", even though that person has nothing to do with "Cornish" (apart
form enjoying a pasty from time to time).

This change filters on the content type of people.person, meaning we'll
only ever get actual other names for the person, not random objects.