Add materialized view for memberships
This adds a new materialized view for all memberships.

The motivation for this is to speed up queries that span a load of
tables. We need to do this when exporting, e.g CSVs of each candidacy.

Django doesn't support materialized views out of the box, so we make the
ORM work in two steps:

1. Run the SQL to create the view in Postgres. This is written manually
   and needs to be updated manually when the underlying models change.
2. Create a Django model in the normal way and add `managed=False` to
   stop Django trying to make it. This allows us to query the view
   without Django trying to get in the way