Define CSV fields in their own class
CSV fields are defined as a mapping of the CSV column name and
a Django ORM database expression.

The idea is that each field knows how to fetch itself from the database, and
all fetching is done in the SQL query at the database end.

This allows the resulting query to be returned to Django without the need to
iterate over each row and perform any sort of action on it.

Fields marked as `core` will _always_ be included in the CSV output.

All other fields are optional.