Don't upload to S3 before starting Textract
This change is because in production (or a production-like environment)
the SOPNs are already uploaded to S3. Uploading them to a new bucket is
just duplication and adds complexity.

What this doesn't solve is how to test uploads in bulk using the sopn
testing tooling, however we can add this back in again if we need it.

The option to bass the bucket name into `TextractSOPNHelper` should
allow for more testing and bulk updating in a later commit