Move to using Textractor for parsing raw responses into Pandas
AWS Maintain a Python library called Textractor for parsing the response
rom Textract.

It manages all the edge cases that we're not going to have thougth of,
and has a nice interface to go with it.

This change moves away from trying to hand parse the JSON response to
relying on AWS's code.