Moderated image matches saved image
This is an improvement on the current code which presents
the user with a preview of an image that matches the
uploaded image, but then saves a different version
(one with a distorted background) following approval.

With this change, the user moderated image matches the
saved image, however neither match the uploaded image.